Added test.unit package, renamed test.integration package, added AssessmentDAOTest and BatchDAOTest
