+++ caliber/src/test/java/com/revature/caliber/test/integration/PDFServiceTest.java
@@ -1,4 +1,4 @@
-package com.revature.caliber.test;
+package com.revature.caliber.test.integration;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
