+++ caliber/src/test/java/com/revature/caliber/test/unit/BatchDAOTest.java
@@ -1,0 +1,7 @@
+package com.revature.caliber.test.unit;
+
+import com.revature.caliber.CaliberTest;
+
+public class BatchDAOTest extends CaliberTest{
+
+}
